+++ HISTORY.md
@@ -8,7 +8,7 @@
 * [ADD] Added the ability to disable an execution during Maven build (Thanks to  Jonathan Piron)
 * [FIX] On Windows, parent directories are now created automatically when adding files to the data archive
 * [CHG] Links are now symbolic by default
-
+* [CHG] Alpha, Beta and RC versions are transformed to a package version ordered before the final release (ex: 1.0~RC1)
 
 ## Version 1.0.1, released 28.02.2013